requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
curl_cffi==0.11.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
frozendict==2.4.6
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.2.6
openai==1.79.0
pandas==2.2.3
peewee==3.18.1
platformdirs==4.3.8
priority==2.0.0
protobuf==6.31.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yfinance==0.2.61
openai>=1.0.0
